+++ modules/beta-public-cluster-update-variant/cluster.tf
@@ -231,6 +231,13 @@ resource "google_container_cluster" "primary" {
         exclusion_name = maintenance_exclusion.value.name
         start_time     = maintenance_exclusion.value.start_time
         end_time       = maintenance_exclusion.value.end_time
+
+        dynamic "exclusion_options" {
+          for_each = maintenance_exclusion.value.exclusion_scope == null ? [] : [maintenance_exclusion.value.exclusion_scope]
+          content {
+            scope = exclusion_options.value
+          }
+        }
       }
     }
   }
@@ -240,9 +247,9 @@ resource "google_container_cluster" "primary" {
   }
 
   timeouts {
-    create = "45m"
-    update = "45m"
-    delete = "45m"
+    create = lookup(var.timeouts, "create", "45m")
+    update = lookup(var.timeouts, "update", "45m")
+    delete = lookup(var.timeouts, "delete", "45m")
   }
   node_pool {
     name               = "default-pool"
@@ -336,6 +343,7 @@ resource "google_container_cluster" "primary" {
       security_group = authenticator_groups_config.value.security_group
     }
   }
+
   notification_config {
     pubsub {
       enabled = var.notification_config_topic != "" ? true : false
@@ -359,6 +367,7 @@ locals {
     "spot",
     "service_account",
     "enable_gcfs",
+    "enable_secure_boot",
   ]
 }
 
@@ -444,7 +453,7 @@ resource "google_container_node_pool" "pools" {
 
   cluster = google_container_cluster.primary.name
 
-  version = lookup(each.value, "auto_upgrade", false) ? "" : lookup(
+  version = lookup(each.value, "auto_upgrade", local.default_auto_upgrade) ? "" : lookup(
     each.value,
     "version",
     google_container_cluster.primary.min_master_version,
@@ -621,8 +630,8 @@ resource "google_container_node_pool" "pools" {
   }
 
   timeouts {
-    create = "45m"
-    update = "45m"
-    delete = "45m"
+    create = lookup(var.timeouts, "create", "45m")
+    update = lookup(var.timeouts, "update", "45m")
+    delete = lookup(var.timeouts, "delete", "45m")
   }
 }